Bugfix for matcherDefinitions - fix indentation
Previously the fflib_MatcherDefinitions.SObjectsWith.matches method, only validated the first record in the given list of records.
Then this first record was matching it return true, without even checking the other records in the list.

Without the change to fflib_MatcherDefinitions.SObjectsWith.matches the unit test with the failingFieldValues was just passing when you modify the fields on the second record.
Now I have added an assertion to assert a failing scenario